Fix a ra_svn deadlock caused by the socket timeout handling introduced for zero-copy server option.  In particular, it interfered with "never block" and "always block" behavior that some parts of the protocol seem to rely on.
We now rely on the network stack to detect stale clients.  This not
worse than before: A server thread could always hang for a long time
waiting for socket to either fail or be serviced.  The only difference
with the zero-copy code is that during this waiting period, other
threads cannot write to one segment of the cache, i.e. there is some
performance hit.

* subversion/libsvn_ra_svn/marshal.c
  (ZERO_COPY_TIMEOUT): drop
  (get_timeout): revert to 1.7.x mechanics

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1465622 13f79535-47bb-0310-9956-ffa450edef68