JavaHL: fix a thread safety bug that has been observed in Subclipse.
The Java SimpleDateFormat class is not thread safe. Solved by synchronizing
the use of the object. Also audited all of our usage of this class and only
this one usage has an issue. There are alternate implementations we could use,
but we would have to copy in the class to our package. I chose to take the simple
approach.

One example we could use if someone objects to using synchronized.

https://code.google.com/p/safe-simple-date-format/

[ in subversion/bindings/javahl/ ]

* src/org/apache/subversion/javahl/types/LogDate.java
  Synchronize the usage of SimpleDateFormat

* src/org/tigris/subversion/javahl/LogDate.java
  Synchronize the usage of SimpleDateFormat


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1476359 13f79535-47bb-0310-9956-ffa450edef68