[Reverted in r1597989]
Add flexible, backward compatible move support to svn_fs_paths_changed.

We introduce a new enum type, svn_move_behavior_t, that will allow us
to specify which changes shall be reported / treated as moves.  For now,
we use it with the log APIs but it is also applicable to merge-related
functionality.

* subversion/include/svn_types.h
  (svn_move_behavior_t): declare new enum type

* subversion/include/svn_fs.h
  (svn_fs_paths_changed3): bumped API with additional move_behavior option
  (svn_fs_paths_changed2): deprecate

* subversion/libsvn_fs/fs-loader.c
  (is_deletion,
   turn_moves_into_copies,
   turn_unique_copies_into_moves): new utility functions for the new
                                   svn_fs_paths_changed3 modes
  (svn_fs_paths_changed3): implement bumped API
  (svn_fs_paths_changed2): implement in terms of the new API

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1525448 13f79535-47bb-0310-9956-ffa450edef68