Better detection of the warning attribute for SVN_EXPERIMENTAL.
First of all we should try to use the __has_attribute macro if available.
Which should allow clang to start using the warning attribute if/when
they add it.  

If not then we'll need to fall back on checking the compiler.  If the compiler
is the gcc driver for clang then it doesn't support __has_attribute and it
also won't set __clang__.  So we need to check for __llvm__ to properly
exclude this.  There's no way to gracefully start using the attribute
if the gcc driver gains support for this, but that's basically gcc's fault
for not having feature detection functionality.

* subversion/include/svn_types.h
  (SVN_EXPERIMENTAL): Use __has_attribute if available and __llvm__ to
    exclude the gcc driver for clang.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1532954 13f79535-47bb-0310-9956-ffa450edef68