Fix a commit problem for copies of mixed revision trees, by properly detecting when 'adds' on top of a not present node should really be committed as replacement, just like how we already handled unmodified not present descendants.
The specific testcase in op-depth-tests didn't trigger the automatic
replacement behavior in libsvn_repos/commit.c add_file_or_directory() as
in this case the new node was not a copy (which has this behavior),
but a plain add (which has not).

This fix makes the client commit both the copy and the plain add case
explicitly without relying on the automatic replacement behavior.

* subversion/libsvn_client/commit_util.c
  (handle_descendants): Properly handle not-present descendants of a copy that
    already have some other committable modification registered.

* subversion/tests/cmdline/merge_tests.py
  (merge_binary_with_common_ancestry): Update expected commit notification for
    a mixed revision copy. (The add was already committed as a replacement by
    the magic in libsvn_repos/commit.c)

* subversion/tests/libsvn_wc/op-depth-test.c
  (test_funcs): Remove XFail marker from copy_mixed_rev_mods.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1536931 13f79535-47bb-0310-9956-ffa450edef68