Alter dav_svn_split_uri() so repos_path are fspaths with a leading slash.
This effectively reverts r1537700, but this solution can't be backported
since it may break users of dav_svn_split_uri() that assume it does not
provide paths with leading slashes.  mod_dontdothat made this assumption
so it's not unlikely if someone has written a 3rd party module for httpd
that it does so as well.

While the repos_path will have a leading slash it still isn't necessarily
canonical.

* subversion/mod_dav_svn/repos.c
  (dav_svn_split_uri): When copying into the repos_path step back one
    character to include the slash.

* subversion/mod_dav_svn/mod_dav_svn.c
  (dav_svn__translate_name): Revert r1537700 changes to this function.

* tools/server-side/mod_dontdothat/mod_dontdothat.c
  (is_this_legal): Canonicalize the path instead of just assuming that
    taking a leading slash is what is needed.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1537812 13f79535-47bb-0310-9956-ffa450edef68