Improve consistency of error reporting in 'svn' and 'svnadmin', mainly by channeling nearly all errors thrown within sub_main() up to a single handler in main().
This is a source code improvement: no functional change.

This is a follow-up to r1368252 which initially separated the initilization
and tear-down into a separate function from the bulk of the logic in these
two programs. It was my intention then to follow up with something like
this. Similar factorization of the other command-line tools will follow.

* subversion/svn/svn.c,
  subversion/svnadmin/svnadmin.c
  (EXIT_ERROR, SVN_INT_ERR): Remove.
  (sub_main): Let this return a standard Subversion error object, and add an
    exit_code output parameter for those cases where a non-zero exit code is
    needed without an error message.
  (main): Display any returned error here.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1543805 13f79535-47bb-0310-9956-ffa450edef68