Provide an error rather than crashing if the client provides no auth_baton.
This has always been really unfriendly to anyone trying to write a client for
the first time.  Most of the bindings help you by setting an empty auth_baton,
but for instance Python doesn't and you'll just get a segfault.

We could check before calling these functions but then we'd litter the code
with these checks so this seems like a reasonable alternative.  The set call
doesn't end up doing what it says but it provides no way to return an error
so the error will just have to wait until we try to get credentials.

* subversion/libsvn_subr/auth.c
  (svn_auth_set_parameter, svn_auth_get_parameter): Check that auth_baton is
    not NULL before trying to use it.
  (svn_auth_first_credentials): Return an error if the auth_baton is empty.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1544320 13f79535-47bb-0310-9956-ffa450edef68