On Windows Stop using FlushFileBuffers() on temporary files that are used for representing symlinks.
Calling FlushFileBuffers() does not always work correctly on network
filesystems, and the atomicity for this specific scenario is already
guaranteed via the rename after the write and the workqueue (only caller).

This fixes checking out of ^/subversion/trunk to a NAS for me, where it used
to fail because we have a symlink for svnpubsub.

* subversion/libsvn_subr/subst.c
  (create_special_file_from_stream): Stop using svn_io_write_unique().


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1547774 13f79535-47bb-0310-9956-ffa450edef68