backport.pl: Avoid problems like Bert fixed in r1567239.
That problem was created by a revision not being prefixed by r.  So don't
require that.

Additional unreliability existed since the code used the r prefix to decide
if a line was a revision line and the revision list can wrap to the next line.
However, the logsummary line is difficult to distinguish from a wrapped revision
list.  So now we only treat a line as a revision list if all characters are used
up in calculating the revision list.  This also means if the first line used as
a revision list doesn't match the pattern then no revisions will be matched and
the line will error about having neither revisions or branch.  There is still
one ambiguous case.  If a revision list wraps and has some extra garbage in it,
it'll be treated as the log summary.  Can't really fix that without changing
the format.

* tools/dist/backport.pl
  (parse_entry): Revision lines match all characters following the repeating
    pattern we're using.  The r prefix is optional.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1567519 13f79535-47bb-0310-9956-ffa450edef68