backport.pl interactive mode: Grow a "show full entry" feature.
The default prompt includes only the revision numbers, logsummary, and votes.   
The full entry display will show Justification and Notes as well.  I expect
this will be useful to get more context on an entry prior to diffing, merging,
or voting on it.  (A "run 'svn log' on this entry" mode already exists.)

* tools/dist/backport.pl
  (handle_entry): Add 'v' mode for 'View entry' (or 'Verbose listing').
  (backport_usage): Document new feature.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1567577 13f79535-47bb-0310-9956-ffa450edef68