Fix a pool usage bug in the svn_rangelist_inheritable2() and consequently svn_mergeinfo_inheritable2() APIs. The result was not completely allocated in the result pool.
As far as I can tell the current trunk Subversion code does not rely on the
output being allocated in the result pool. I found this by code inspection.

This bug was introduced in r866302.

Note also that this bug was hiding a bug in the previous line which I fixed
a few minutes ago in r1569731.

* subversion/libsvn_subr/mergeinfo.c
  (svn_rangelist_inheritable2): Duplicate the remaining ranges into the
    result pool.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1569764 13f79535-47bb-0310-9956-ffa450edef68