Remove the overthought svn_fs_base__key_compare() function.
The old implmenetation ran strlen() on both strings, which in my opinion is a
really dubious optimization.  I kinda doubt that walking both strings twice in
cases where the key is the same size is optimally better than saving the
character comparisons when they are different sizes.  I'd bet that we're
comparing identically sized strings a lot of the time.  Forcing -1, 0, or 1
as the only return values seems entirely pointless to me since the only thing
we ever do is test for 0 when calling this.

This incidentally fixes a compiler warning since we we were putting the return
of strlen into an int which isn't necessarily the same size as the strlen().

* subversion/libsvn_fs_base/key-gen.h
  (svn_fs_base__key_compare): Convert the function to a macro to strcmp and
    update the documentation to not promise things strcmp() doesn't promise.

* subversion/libsvn_fs_base/key-gen.c
  (svn_fs_base__key_compare): Delete.



git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1570706 13f79535-47bb-0310-9956-ffa450edef68