Resolve an assertion by properly calculating a repository path. For this specific delete case it calculates the path of what would have been added if it wasn't directly deleted... Yuck.
An alternative way to resolve the problem would be to remove the
assumption that this case has a path... which appears simpler, but
makes things very hard in the commit logic.

I don't mark the test fixed with this commit as this only resolves
the assertion... The commit now fails with an 'out of date' error,
while it should provide an error about the ancestor not being committed.

* subversion/libsvn_wc/status.c
  (get_repos_root_url_relpath): Fix repository path calculations for
    working-deletes.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1571787 13f79535-47bb-0310-9956-ffa450edef68